ecommerce/requirements.txt
astroid==1.5.3
attrs==17.3.0
boltons==16.1.1
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.11.8
django-debug-toolbar==1.9.1
django-debug-toolbar-request-history==0.0.6
django-extensions==1.9.8
django-js-reverse==0.7.3
django-model-utils==3.1.2
django-naomi==0.8
django-rest-swagger==2.1.2
djangorestframework==3.7.3
dodgy==0.1.9
factory-boy==2.11.1
Faker==0.8.16
fixmydjango==0.3.1
idna==2.6
invoke==0.22.0
isort==4.2.15
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
openapi-codec==1.3.2
pep8-naming==0.4.1
Pillow==5.1.0
pluggy==0.6.0
prospector==0.12.7
psycopg2==2.7.3.2
py==1.5.2
pycodestyle==2.0.0
pydocstyle==2.1.1
pyflakes==1.6.0
pylint==1.7.4
pylint-celery==0.3
pylint-common==0.2.5
pylint-django==0.7.2
pylint-flask==0.5
pylint-plugin-utils==0.2.6
pytest==3.3.0
pytest-django==3.1.2
python-dateutil==2.7.3
python-decouple==3.1
pytz==2017.3
PyYAML==3.12
requests==2.9.1
requirements-detector==0.5.2
setoptconf==0.2.0
simplejson==3.13.2
six==1.11.0
snowballstemmer==1.2.1
sqlparse==0.2.4
termcolor==1.1.0
text-unidecode==1.2
typing==3.6.2
Unipath==1.1
uritemplate==3.0.0
urllib3==1.22
wrapt==1.10.11
